requirements.txt
numpy=1.23.5
pandas=2.0.3
scikit-learn=1.3.2
scipy=1.10.1
torch=2.2.2
torchvision=0.17.2
pillow=10.2.0
d2l=1.0.3
